An upgrade of the PostLexicalAnalyzer to 1.1 Flite functionality.
The fix_apostrophe_s code has been upgraded to also deal with "'ve",
"'ll" and "'d" suffixes that follow a vowel. It will prepend the schwa
before these as well. However, since the tokenizer has not yet been
upgraded to deal with those, I can't test this code yet.

Also, a post lexical routine has been added that converts the
pronunciation of "the" from "d ax" to "d iy" if the following word
starts with a vowel.