There were a couple of off-by-one errors in the ClusterUnitSelector that were causing the utteranceTimeTest regression test to fail. Fixed those.
The default initial F0 and F0 range for the ClusterUnitVoice has changed
in Flite. Changed CMUClusterUnitVoice to match that for FreeTTS.

After this checkin all utterance regression tests pass. WOOHOO!

Next up are the wave output tests. (these may be trickier due to the
fixed point math changes in Flite 1.1)

.