This is absolutely the last commit before the code freeze. The main change is in FreeTTSEngineCentral.java.
It used to be the case that JSAPI modes with no valid voices is
not added to the private list of modes in FreeTTSEngineCentral. This
creates a few problems. Suppose a user installed MBROLA at the
wrong location. This causes the mbrola mode not to be added to the
private list of modes in FreeTTSEngineCentral. If that user then
try to run an MBROLA demo, he will not get any MBROLA validity messages,
because the mbrola mode is not in the list, and no mbrola validity
check will be done.

To alleviate this problem, the previous version
would check for validity when loading the synthesizer.properties file,
and print out an error messages say that MBROLA is not found. However,
that itself also creates a problem - someone who does not use MBROLA
will get that message just by running any of the JSAPI demos.

So, now, the best solution would be:

1) Load all the modes, and add all of them to the private list. Do not
   print any validity error message when loading.
2) In the createEngineList() method, if the given ModeDesc matches a
   mode in the private list, check for validity. If its not valid,
   print out an error message.

In short, the outcome:

1) A user who does not use MBROLA will not get any MBROLA validity messages.
   This is because in createEngineList(), the required ModeDesc will
   not match MBROLA, so no validity check will be done on MBROLA mode.

2) An MBROLA user will get the appropriate validity messages, since his
   required ModeDesc will match MBROLA modeDesc, and validity check will be
   done.