isAWTRunning was generating an AccessControlException when running from an applet. This was due to isAWTRunning attempting to get the ThreadGroup of its parent.  Applets are not allowed to get the ThreadGroup of their parents.
To work around this, isAWTRunning now catches the
AccessControlException and if so assumes that we are running from an
applet, in which case AWT must be running, so it returns true.

I'm not sure if this is indeed the best behavior, perhaps there is
another way to check to see if AWT is running that is applet safe.

Any ideas would be appreciated.