Update addenda.  Add some more info that the FestVoxToFreeTTS stuff is really good for en_US, but much harder for other locales.  Update build.xml to package stuff more like sphinx-4.
NOTE: I was finally able to import the CMU ARCTIC voices (the main
problem was the Scheme-based sort, so I resorted to a hacky method
of using Flite's tools for some of the migration, and FreeTTS for
the rest).  It took an amazing 12 hours to run on my Mac.  In any
case, it still isn't working well because the unit selection method
depends upon different features.  So...I'll take a poke at writing
a custom unit selection mechanism for these voices.